Add files via upload
With this program, i can broadcast with many people using the same program by typing their name in the GUI section, using the same router. I did this program on CMP 405. And was my nine homework. fooling homework six. Then the other person can reply back to me as well. And can look for me in the too.